Split boxref to separate file
